Add v8_set_flags to core
