+++ core/lib.rs
@@ -1,3 +1,4 @@
+// Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
 #[macro_use]
 extern crate log;
 extern crate futures;
@@ -6,8 +7,11 @@ extern crate libc;
 mod isolate;
 mod js_errors;
 mod libdeno;
+mod flags;
 
 pub use crate::isolate::*;
 pub use crate::js_errors::*;
 pub use crate::libdeno::deno_buf;
 pub use crate::libdeno::deno_mod;
+pub use crate::flags::v8_set_flags;
+
